refactor(shared): extract IsInlineSafeMime + BuildContentDisposition к shared/http/headers (ADR-4 close)
Closes v0.161.0 ADR-4 carry-forward (Tier 1, deferred per senior
enterprise approach for the security-focused hotfix). The helpers
previously lived в `internal/modules/documents/interfaces/http/handlers/`
и были duplicated conceptually в files (which had no clickjacking
protection at all). One authoritative copy в shared eliminates drift и
makes the v0.156.0 ADR-2 (inline whitelist) + ADR-3 (RFC 5987 filename
encoding) hardening available для any module serving downloadable
artifacts.

New package: `internal/shared/infrastructure/http/headers/`
- `inline_mime.go` — IsInlineSafeMime(mime string) bool — whitelist
- `inline_mime_test.go` — 17 cases (preview-safe + executable rejects
  + charset-suffix tolerance + case-insensitivity)
- `content_disposition.go` — BuildContentDisposition + stripControlBytes
- `content_disposition_test.go` — 7 cases (ASCII / cyrillic RFC 5987 /
  CRLF injection / control-byte rejection / empty fallback)

Updated callers:
- `documents/interfaces/http/handlers/document_handler.go` — import
  shared pkg + 2 call sites qualified (`headers.IsInlineSafeMime`,
  `headers.BuildContentDisposition`)

Files DownloadFile currently returns presigned URL JSON (no byte
streaming through handler), so Content-Disposition header is set by
MinIO/S3 on the actual GET. Shared helpers ready when/if files API
restructures к stream-through — out of scope for this polish.

Deleted: 4 documents helper files (2 .go + 2 _test.go, copied verbatim
к shared).

Pure refactor — no behavior change. All tests pass (shared headers /
documents handlers / files handlers).

Refs #317
